requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
albucore==0.0.21
albumentations==1.4.7
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.1.0
Authlib==1.3.0
auto_gptq==0.7.1+rocm5.7.3
autoawq==0.2.5
autoawq_kernels==0.0.9
autotrain-advanced==0.8.10
# bitsandbytes @ file:///mnt/additional/projects/sLLM/bitsandbytes
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
codecarbon==2.3.5
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cryptography==42.0.5
cycler==0.12.1
datasets==2.19.1
diffusers==0.27.2
dill==0.3.8
dnspython==2.7.0
docstring_parser==0.16
einops==0.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
evaluate==0.4.1
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.7
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.3.1
gekko==1.2.1
gguf==0.14.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
hf_transfer==0.1.9
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
inflate64==1.0.1
iniconfig==2.0.0
invisible-watermark==0.2.0
ipadic==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiwer==3.0.3
joblib==1.4.0
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.12.0
lion-pytorch==0.2.3
loguru==0.7.2
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
networkx==3.4.2
ninja==1.11.1.3
nltk==3.8.1
numpy==1.26.4
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.25.1
nvitop==1.3.2
onnx==1.17.0
onnxruntime_extensions==0.14.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
# optimum==1.24.0
# Editable install with no version control (optimum-amd==0.2.dev0)
# -e /mnt/additional/projects/sLLM/optimum-amd
optuna==3.6.1
orjson==3.10.15
packaging==24.0
pandas==2.2.2
peft==0.12.0
pillow==10.3.0
pluggy==1.5.0
pretty-errors==1.2.25
prometheus_client==0.21.1
propcache==0.3.0
protobuf==4.23.4
psutil==7.0.0
py-cpuinfo==9.0.0
py7zr==0.21.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pybcj==1.0.3
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.19.1
pyngrok==7.1.6
pynvml==12.0.0
pyparsing==3.2.1
pyppmd==1.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-triton-rocm==3.2.0+git4b3bb1f8
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.1
pyzstd==0.16.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.31.0
responses==0.18.0
rich==13.9.4
rich-toolkit==0.13.2
rouge==1.0.1
rouge-score==0.1.2
sacremoses==0.1.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.5.0
scipy==1.14.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
shtab==1.7.1
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.37.2
stringzilla==3.12.2
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.5.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.6.0
timm==1.0.7
tokenizers==0.19.1
tomli==2.2.1
torch==2.7.0.dev20250306+rocm6.3
torchaudio==2.6.0.dev20250306+rocm6.3
torchmetrics==1.4.0
torchvision==0.22.0.dev20250306+rocm6.3
tqdm==4.66.4
transformers==4.44.0
triton==3.2.0
trl==0.9.6
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.9.16
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
Werkzeug==3.0.2
xgboost==2.0.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
